Remove panel from view when there are no entries

+++ app/Http/Controllers/MonthlyTrackingController.php
@@ -86,12 +86,14 @@ class MonthlyTrackingController extends Controller
 
       if($request->has('month_id')) {
         $trackedMonth = TrackedMonth::where('id', $request->get('month_id'))->first();
+        if($trackedMonth) {
+          $trackedMonthStartDate = Carbon::createFromDate($trackedMonth->year, $trackedMonth->month)->startOfMonth();
+          $trackedMonthEndDate = Carbon::createFromDate($trackedMonth->year, $trackedMonth->month)->endOfMonth();
+          if(!$date->between($trackedMonthStartDate, $trackedMonthEndDate)){
+            // updated date falls outside of the current tracked month attached to the updated entry so it needs to be moved to the correct month
+            $trackedMonth = TrackedMonth::where('user_id', Auth::user()->id)->where('month', $date->format('m'))->where('year', $date->format('Y'))->first();
+          }
 
-        $trackedMonthStartDate = Carbon::createFromDate($trackedMonth->year, $trackedMonth->month)->startOfMonth();
-        $trackedMonthEndDate = Carbon::createFromDate($trackedMonth->year, $trackedMonth->month)->endOfMonth();
-        if(!$date->between($trackedMonthStartDate, $trackedMonthEndDate)){
-          // updated date falls outside of the current tracked month attached to the updated entry so it needs to be moved to the correct month
-          $trackedMonth = TrackedMonth::where('user_id', Auth::user()->id)->where('month', $date->format('m'))->where('year', $date->format('Y'))->first();
         }
       } else {
         $trackedMonth = TrackedMonth::where('user_id', Auth::user()->id)->where('month', $date->format('m'))->where('year', $date->format('Y'))->first();
@@ -137,9 +139,20 @@ class MonthlyTrackingController extends Controller
       }
 
       $record->save();
-      $request->session()->put('selectedMonth', $trackedMonth->month);
-      $request->session()->put('selectedYear', $trackedMonth->year);
-      $request->session()->put('saved', $record->occurred_at);
+
+      //check if there are any remaining records for the iltracked month, if not delete it.
+      $monthRecords = MonthlyTrackingRecord::where('month_id', $request->get('month_id'))->get();
+      if($monthRecords->isEmpty()){
+        $this->deleteMonth($request->get('month_id'));
+        $request->session()->forget('selectedMonth');
+        $request->session()->forget('selectedYear');
+        $request->session()->forget('saved');
+      } else {
+        $request->session()->put('selectedMonth', $trackedMonth->month);
+        $request->session()->put('selectedYear', $trackedMonth->year);
+        $request->session()->put('saved', $record->occurred_at);
+      }
+
 
       if($oldCategory && $oldCategory != $record->category){
         $this->checkCategory(Auth::user()->id, $oldCategory);
@@ -162,7 +175,7 @@ class MonthlyTrackingController extends Controller
           '01' => 'January'
         ]
       ];
-      
+
       return json_encode($data);
     }
 
@@ -227,13 +240,14 @@ class MonthlyTrackingController extends Controller
 
     function deleteMonth($id)
     {
-      $record = TrackedMonth::findOrFail($id);
+      $record = TrackedMonth::find($id);
+      if($record) {
+        if($record->user_id && $record->user_id != Auth::user()->id) {
+          abort(403, 'Unauthorized action.');
+        }
 
-      if($record->user_id && $record->user_id != Auth::user()->id) {
-        abort(403, 'Unauthorized action.');
+        $record->delete();
       }
-
-      $record->delete();
 
       return;
     }
